Added More Fuels and Checking for Valid Container
Now I believe all valid fuels are allowed in the Furnace via
infinitechests.

Also upon a sign change (Creating the infinitechests sign) it will check
the container. If there is a valid container the sign will be created,
if there is not a supported container or no container the sign will pop.